refactor: consolidate Deck import types
