+++ src/features/deck-import/ui/DeckImportView.spec.tsx
@@ -3,7 +3,8 @@ import userEvent from "@testing-library/user-event";
 import "@testing-library/jest-dom/vitest";
 import { describe, expect, it, vi } from "vitest";
 
-import type { DeckImportPreview, DeckImportResult } from "../model/deckImportTypes";
+import type { DeckImportPreview } from "../hooks/useDeckImport";
+import type { DeckImportResult } from "../model/deckImportExecution";
 
 import { DeckImportView } from "./DeckImportView";
