+++ src/features/deck-import/ui/DeckImportView.tsx
@@ -4,7 +4,8 @@ import { AiOutlineCloudDownload } from "react-icons/ai";
 import { Button } from "@/shared/ui/button";
 import { Code, Description } from "@/shared/ui/content";
 import { Upload } from "@/shared/ui/forms";
-import type { DeckImportPreview, DeckImportResult, DeckImportStorageMode } from "../model/deckImportTypes";
+import type { DeckImportPreview } from "../hooks/useDeckImport";
+import type { DeckImportResult, DeckImportStorageMode } from "../model/deckImportExecution";
 
 interface DeckImportViewProps {
   onChange?: (file: File) => void;
